- Fixed an app breaking bug - View description for details
Around the 100 hour mark, the website changes how it displays information
We previously had a subTimer which would contain up to 24 hours, and a mainTimer which contains days

Now the website only populates the mainTimer section with the hours remaining in FULL -- i.e 78 hours instead of 3 days 6 hours
This caused the string to be empty as we were now checking the incorrect fields. (to fix this I added a check to see if the response is null or not)

We also had an issue with our calculation which was then taking the MainTimer * 24 and adding this total onto the days, but since this is now automatically done we would get a response of double the remaining time.  Fixed this by null checking the string once again and ignoring the calculation if it is empty.

---
This should now all be fixed, however I am unable to test if it is bug free until the website updates back to the original display method (or at least cannot be bothered to check)